backend/requirements.txt
Flask==3.0.0
Flask-CORS==4.0.0
PyPDF2==3.0.1
sentence-transformers==2.2.2
faiss-cpu==1.7.4
numpy==1.24.3
werkzeug==3.0.1
